(CRITICAL) Modified String Resource File
Modified the content of the String Resource file by using the extracted string values from the existent Activities as its resources. The string values are further organized into global and class_specific categories to facilitate readability and maintainability. This is implemented to allow translations and reusability for future development efforts or use cases